+++ load-test/bin/redis-create-prepared-checks.js
@@ -3,7 +3,7 @@
 'use strict'
 
 require('dotenv').config()
-const redis = require('../redis.service')
+const redis = require('../services/redis.service')
 const { performance } = require('perf_hooks')
 const exampleCheck = require('../pupil-api/preparedCheck')
